Give assistant file links a home
