+++ packages/app/src/utils/inline-path.ts
@@ -7,8 +7,6 @@ export interface InlinePathTarget {
   lineEnd?: number;
 }
 
-export type OpenFileDisposition = "main" | "side";
-
 const FILE_PROTOCOL = "file:";
 const INLINE_LINE_FRAGMENT = /^L([0-9]+)(?:-L?([0-9]+))?$/i;
 const ASSISTANT_FILE_EXTENSIONS = new Set([
