+++ packages/app/src/utils/workspace-file-open.ts
@@ -1,0 +1,1 @@
+export type OpenFileDisposition = "main" | "side";
